move swc config to top level
